[CBR-165] Add QuickCheck round-up property on Mnemonics package
As part of my investigation, I suspected the `fromMnemonic` to be
erroneous being guilty in the wallet restoration issue (from a bug
discovered in the original implementation related to the checksum
verification). However, it doesn't seem to be wrong as I couldn't
find any counter examples with QC. It seems that we always get back
the same seed from an encoded mnemonic sentence.